refactor: remove ContactInfoItemSkeleton and SidePanelContactInfoItem, introduce SidePanelHeaderInfoItem for improved structure and interactivity
